Refactor to get_releases function
